+++ src/abacus/protocol_version.cpp
@@ -1,0 +1,18 @@
+// Copyright (c) Steinwurf ApS 2020.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#include "protocol_version.hpp"
+#include "version.hpp"
+
+namespace abacus
+{
+inline namespace STEINWURF_ABACUS_VERSION
+{
+uint8_t protocol_version()
+{
+    return 0;
+}
+}
+}