+++ src/abacus/protocol_version.hpp
@@ -1,0 +1,19 @@
+// Copyright (c) Steinwurf ApS 2020.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+#pragma once
+
+#include <cstdint>
+
+#include "version.hpp"
+
+namespace abacus
+{
+inline namespace STEINWURF_ABACUS_VERSION
+{
+/// @return The version of the protocol
+uint8_t protocol_version();
+}
+}